feat: 설문조사 제출 시, 최초 recommend list 생성하도록 보완

+++ src/main/java/com/autoever/carstore/car/dao/CarModelRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CarModelRepository extends JpaRepository<CarModelEntity, Long> {
     @Query("SELECT c FROM CarModelEntity c WHERE c.carModelId = :carModelId")
-    CarModelEntity findByCarModelId(Integer carModelId);
+    CarModelEntity findByCarModelId(Long carModelId);
 }
